Alignment fixes made 64-bit safe
These patches to Jimmy's alignment patch are intended to make
his alignment fixes 64-bit safe, since they are broken on
64-bit architectures.

For future reference, I think we can ensure architecture-independence
by:

1) Casting pointers to unsigned long, not unsigned int. On 64-bit
   architectures an int is 32-bits, and a long is 64-bits.

2) When performing bitwise logic on pointer values, use (~x) to
   define constants rather than the 0xfffff... form, which inherently
   creates a 32-bit mask that will truncate the pointer and crash your
   code.

+++ rddma_src.c
@@ -136,15 +136,17 @@ struct kobj_type rddma_src_type = {
 struct rddma_src *new_rddma_src(struct rddma_dst *parent, struct rddma_bind_param *desc)
 {
 	struct rddma_src *new = kzalloc(sizeof(struct rddma_src) + RDDMA_DESC_ALIGN - 1, GFP_KERNEL);
-    
+
 	if (NULL == new)
 		goto out;
 
 	/* DMA descriptors are embedded in the rddma_src struct, so
 	 * align the struct to what the DMA hardware requires
+	 *
+	 * TJA: made 64-bit safe [use long, not int, to avoid truncating 64-bit pointers]
 	 */
-	if ((unsigned int) new & (RDDMA_DESC_ALIGN-1))
-		new = (struct rddma_src *) (((unsigned int) new + RDDMA_DESC_ALIGN) & ~(RDDMA_DESC_ALIGN - 1));
+	if ((unsigned long) new & (RDDMA_DESC_ALIGN-1))
+		new = (struct rddma_src *) (((unsigned long) new + RDDMA_DESC_ALIGN) & ~(RDDMA_DESC_ALIGN - 1));
 	rddma_clone_bind(&new->desc, desc);
 	new->kobj.ktype = &rddma_src_type;
 	kobject_set_name(&new->kobj,"%s#%llx:%x",new->desc.src.name, new->desc.src.offset, new->desc.src.extent);
